pyproject.toml
[tool.poetry]
name = "cncrd-ext-audio"
version = "3.0.0"
description = "Audio-related functionality extension for Concord"
authors = ["Nariman Safiulin <woofilee@gmail.com>"]
license = "MIT"

readme = "README.md"
repository = "https://github.com/Roolat/concord-ext-audio"

keywords = ["discord", "concord", "bot", "audio"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Natural Language :: Russian",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

packages = [
    { include = "concord" },
]

[tool.poetry.dependencies]
python = "^3.6"
cncrd = "^1.0.0"
"discord.py" = { git = "https://github.com/Rapptz/discord.py.git", branch = "rewrite", extras = ["voice"] }

[tool.poetry.dev-dependencies]
black = "^18.9b0"
pre-commit = "^1.10"

[tool.black]
line-length = 80
py36 = true
